Decompose catalog member correspondence
Separate open-signature planning and frozen-context join projection from the public correspondence plan facade without changing its contract.
